requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==25.1.0
Automat==25.4.16
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.16.0
channels==4.2.2
chardet==5.2.0
charset-normalizer==3.4.1
chromedriver-autoinstaller==0.6.4
click==8.1.7
comm==0.2.0
constantly==23.10.4
crispy-bootstrap5==2023.10
cryptography==42.0.4
cssselect==1.3.0
cycler==0.12.1
dacite==1.8.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.7
django-bootstrap3==23.6
django-ckeditor==6.7.2
django-crispy-forms==2.1
django-js-asset==3.0.1
django-model-utils==4.3.1
django-notifications-hq==1.8.3
django-widget-tweaks==1.5.0
django4django==0.0.2
djangorestframework==3.14.0
djangorestframework-xml==2.0.0

ghostscript==0.7
h11==0.14.0
html5lib==1.1
htmlmin==0.1.12
hyperlink==21.0.0
idna==3.10
ImageHash==4.3.1
incremental==24.7.2
ipykernel==6.28.0
ipython==8.19.0
isoduration==20.11.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.1
json5==0.9.14
jsonfield==3.1.0
jsonpointer==2.4
jsonschema==4.20.0



numpy==2.2.3

pillow==11.1.0

Protego==0.4.0
psutil==5.9.7


PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
PythonTurtle==0.3.2
pytweening==1.2.0
pytz==2025.1

pyzmq==25.1.2
queuelib==1.8.0
referencing==0.32.0
requests==2.32.3
requests-file==2.1.0
